Expand converter map
