+++ lib/lib.js
@@ -1,6 +1,6 @@
 export { version } from "../package.json";
 import AsbindInstance from "./asbind-instance/asbind-instance";
-export { converters } from "./asbind-instance/bind-function";
+export { converters } from "./asbind-instance/type-converters";
 
 export async function instantiate(source, importObject) {
   let asbindInstance = new AsbindInstance();
